docs-36

+++ jenkins/api.sh
@@ -1,12 +1,18 @@
 #!/bin/bash -e
 #
+# This script pulls in the apihome repo into the content directory
 #
-#
+if [ -z "${1}" ]
+then
+  echo "Usage: ${0} <branch>"
+  exit 1
+fi
+BRANCH="${1}"
 mkdir -p content
 cd content
 rm -rf api
 git clone git@git.hpcloud.net:DevExDocs/apihome.git
 mv apihome api
 cd api
-git checkout "${1}"
-git pull origin "${1}"
+git checkout "${BRANCH}"
+git pull origin "${BRANCH}"
